+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ## 1.0.5 ()
 
-- add options: show bible verses
+- add options: show inline verse numbers
 
 ## 1.0.4 (12/16/2019)